Refactoring structure of project
