+++ .gitattributes
@@ -1,6 +1,2 @@
-*.tsv filter=lfs diff=lfs merge=lfs -text
-<<<<<<< HEAD
 *.psd filter=lfs diff=lfs merge=lfs -text
 **.tsv filter=lfs diff=lfs merge=lfs -text
-=======
->>>>>>> 6b3b33e941f59cac9e1c951c66b4bf542d8da78d
