--- data/ext/simple/AllGraph_dict.sample_500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ebd8bc974ed7ddc479649127e2b16adbb87e01be32bcfbf19e16b3cac8dc10a5
-size 430072
