--- data/ext/simple/AllGraph_dict.sample_5000.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1c84159358e6735b73ad500c053698c2ec5954de840fc4c4e5c4c6351ebd8794
-size 4850023
