--- data/ext/simple/AllGraph_dict.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:47e99baef5defc9b1c8e5686a9fedcd98c62b07a0b64cab2c2623d82fe1a012e
-size 224067214
