--- data/ext/simple/GeneratedFiles/AllGraph_tfidf.sample_500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e48e89e76abc7fd3d809f42c411b28a928b871218284504898dd559658728720
-size 750703
