--- data/ext/simple/GeneratedFiles/AllGraph_tfidf_candidateLabels.sample_500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:19eff9d91ca81a477ccd3ada316a92f1e6e5081d9af698dabe727f1a8c012dcf
-size 1188
