--- data/ext/simple/cluster_with_externalIds500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f5477eac0263f75d679f256731cfe514e74922567d9effe463c09822785535ae
-size 8003
