--- data/ext/simple/cluster_with_internalID.500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f89b5d8507fd5eb3d6c89ef6b5978e849ffa1a6182752652534909fd2f24e983
-size 7265
