--- data/ext/simple/cluster_with_internalID.sample_500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:28e36d42c034372e69d2c35fe3708c4fbf066ff25422b5eb66fb01a1b939d5be
-size 7263
