--- data/ext/simple/coordinates.sample_500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1b8bca2369c16f2754f87850df1b23fd41d3aaede58c3cabd2f5f55053213e31
-size 16365
