--- data/ext/simple/logs/2017_06_01_15_13_06_126480_412612.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d0670cd31d23611d3df7e3f69cc6f358953c75449c73f75f73f6abd037447f78
-size 19494
