--- data/ext/simple/region_names.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2c6c3dc8a02b44b2f1941b828e55f6a5253775a8781359d5160750d4aee8f247
-size 137
