--- data/ext/simple/tsne_cache.sample_50000.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:fddcd49fe98b75b20d45af2de6598b8f7fb096d8e1f355c20e30d49b99ccc98f
-size 1735631
