--- data/ext/simple/vectors.sample_500.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:625933392c8a31df434c5ee2d38621337b34b796cdadfbbefd308b74012fca2e
-size 1208048
