--- data/ext/simple/vectors.sample_5000.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5badd8f37f02c505c937a2b0656922dd91a91586cb46d6f7d6781e85f6f62e62
-size 12088319
