--- data/ext/simple/vectors.sample_50000.tsv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:663eabc605d2e234fb45494864189f107e0cf076276a033a4ffeb589dd4d5493
-size 120892776
